+++ src/mdast-robust-tables.js
@@ -12,7 +12,7 @@
 /* eslint-disable no-param-reassign */
 import { toHast as md2hast } from 'mdast-util-to-hast';
 import { toHtml as hast2html } from 'hast-util-to-html';
-import { visit } from 'unist-util-visit';
+import { visit, CONTINUE } from 'unist-util-visit';
 
 /**
  * Converts tables to HTML
@@ -23,26 +23,32 @@ import { visit } from 'unist-util-visit';
 export default function robustTables(tree) {
   visit(tree, (node) => {
     if (node.type !== 'table') {
-      return visit.CONTINUE;
+      return CONTINUE;
     }
     let html = '<table>\n';
     (node.children /* c8 ignore next */ || []).forEach((row) => {
       html += '  <tr>\n';
       (row.children /* c8 ignore next */ || []).forEach((cell) => {
-        let align = '';
+        let attrs = '';
         if (cell.align === 'right') {
-          align = ' align="right"';
+          attrs = ' align="right"';
         } else if (cell.align === 'center') {
-          align = ' align="center"';
+          attrs = ' align="center"';
         } else if (cell.align === 'both') {
-          align = ' align="justify"';
+          attrs = ' align="justify"';
         }
         if (cell.valign === 'middle') {
-          align += ' valign="middle"';
+          attrs += ' valign="middle"';
         } else if (cell.valign === 'bottom') {
-          align += ' valign="bottom"';
+          attrs += ' valign="bottom"';
         }
-        html += `    <td${align}>`;
+        if (cell.rowSpan > 1) {
+          attrs += ` rowspan="${cell.rowSpan}"`;
+        }
+        if (cell.colSpan > 1) {
+          attrs += ` colspan="${cell.colSpan}"`;
+        }
+        html += `    <td${attrs}>`;
 
         // if cell contains only 1 single paragraph, unwrap it
         let { children } = cell;
@@ -54,7 +60,8 @@ export default function robustTables(tree) {
           if (child.type === 'html') {
             html += child.value;
           } else {
-            const cellHtml = hast2html(md2hast(child));
+            const hast = md2hast(child, { allowDangerousHtml: true });
+            const cellHtml = hast2html(hast, { allowDangerousHtml: true });
             if (child.type === 'code') {
               // code needs special treatment, otherwise the newlines disappear.
               html += cellHtml.replace(/\r?\n/g, '<br>');
@@ -71,7 +78,7 @@ export default function robustTables(tree) {
     node.type = 'html';
     node.value = html;
     delete node.children;
-    return visit.CONTINUE;
+    return CONTINUE;
   });
   return tree;
 }