+++ src/mdast-sanitize-heading.js
@@ -13,36 +13,78 @@ import { visit } from 'unist-util-visit';
 
 /**
  * Sanitizes headings:
- * - (re)move images
+ * - (re)move images ('before', 'both', 'after')
+ * - converts BREAKs inside headings to <br>.
  *
  * @param {object} tree
+ * @param {object} [opts] options
+ * @param {string} [opts.imageHandling] specifies how images are handled. defaults to 'after'.
  * @returns {object} The modified (original) tree.
  */
-export default function sanitizeHeading(tree) {
+export default function sanitizeHeading(tree, opts = {}) {
+  const { imageHandling = 'after' } = opts;
   visit(tree, (node, index, parent) => {
     const { children: siblings = [] } = parent || {};
     const { children = [] } = node;
+    let after = index + 1;
     if (node.type === 'heading') {
       for (let i = 0; i < children.length; i += 1) {
         const child = children[i];
         if (child.type === 'image') {
-          // move after heading
-          children.splice(i, 1);
-          i -= 1;
           const para = {
             type: 'paragraph',
             children: [child],
           };
-          siblings.splice(index + 1, 0, para);
+          children.splice(i, 1);
+          i -= 1;
+          if ((i < 0 && imageHandling !== 'after') || imageHandling === 'before') {
+            // move before heading
+            siblings.splice(index, 0, para);
+            // eslint-disable-next-line no-param-reassign
+            index += 1;
+            after += 1;
+          } else {
+            // move after heading
+            siblings.splice(after, 0, para);
+            after += 1;
+          }
         }
       }
+
+      // move leading breaks before heading
+      while (children[0]?.type === 'break') {
+        const brk = children.shift();
+        siblings.splice(index, 0, brk);
+        // eslint-disable-next-line no-param-reassign
+        index += 1;
+      }
+      // move trailing breaks after heading
+      let last = children.length - 1;
+      while (children[last]?.type === 'break') {
+        const brk = children.pop();
+        siblings.splice(index + 1, 0, brk);
+        last -= 1;
+      }
+      // convert inline breaks to <br>
+      for (let i = 0; i < children.length; i += 1) {
+        if (children[i].type === 'break') {
+          children[i] = {
+            type: 'html',
+            value: '<br>',
+          };
+        }
+      }
+
       // remove empty headings
       if (!children.length) {
         siblings.splice(index, 1);
-        return index;
+        // eslint-disable-next-line no-param-reassign
+        index -= 1;
       }
+      return index + 1;
     }
-    return visit.CONTINUE;
+
+    return index + 1;
   });
   return tree;
 }